requirements.txt
viam-sdk==0.47.0
typing-extensions
aiofiles>=23.0.0